Add OutputTree, a helper for writing structured data from system tests
Basically this is a Dune::ParameterTree enriched by
- automatic write to file upon destructin the tree object
- setting prefixes for keys to be applied to all operations

The latter is to have dataspecific keys. For example: In a time stepping
loop, the time stepping can set the key prefix. The solver within the
loop then need not know, which key is associated with the timestep, because
the tree object automaticaly prepends the correct tree. Prefixes are handled
stack-like.

This implementation is intended to write our own tests with. In the long
run, it might be a good idea to convince dune-devel of having it in dune-common.