Introduce a special key to customize the filename.
The default behaviour is to generate ini files with
a counting variable. However, a user might want to
control the name of the generated ini file through
the key-dependent value resolution syntax ({}).
This commit introduces the reserved key "__name"
to do so.

TODO: Maybe in the future a combined approach of
numbering and naming is interesting. For now, its
either one or the other.