Add error handling and default behaviour in static_metaini
When calling the python script without having a __STATIC section in the ini file, may happen by accident
the programme exited on a KeyError. Now an error message is thrown to std::err and the script exits (does not output anything to cmake).
When not providing the special keyword __exec_suffix the programme exited on a KeyError. Now it appends an incrementing integer to the target name.