[add_system_test_per_target] Support for user defined external grid files
In case the systemtest uses an external grid file, e.g. a DGF-File the grid has to be avaible
in the build directory. The macro now searches for a folder "grids" in the source directory
and if found copies it to the build directory. The grid files are assumed to be sufficiently small
to not copy huge data.