Handle output if there are no static variations
Until now CMake validated all executable/ini combinations as false and does not build the only executable if
the static section is missing. This exploits the fact that the python script hands over nothing to CMake
in case of no static variations. The output variable remains an empty string.
The only executable has the target basename as name.