Use regular expression to implement escaping even further
Escaping the { and } in key-dependent value lookup was not
possible til now. It is now implemented. Additional, the code
got more robust and pythonic (hopefully). The special key
"__delete" didnt fit this approach. I thought on it again and
came to the conclusion that it is not really necessary after all.
If this is wrong, I can try to implement it again.