[metaini] Manipulate the __name key if --dir is present.
The __name key now always consists of the actual path of the ini file + a unique string without
file extension. This is more useful then having it names after the source patch. This way the __name
key can be used in c++ programmes to generate run-unique output names for output data we want to compare