[command] Make command syntax much more flexible
The commands can now come in any given order. Their order is automatically determined
by the command types. If we have two commands of the same command type the first one is executed first.
This adjusts the expand command and the convergence test command to deal with the new syntax flexibility.
See python test convtest4.mini for an extensive use of the flexible syntax!