Implement a label command.
It takes two arguments: The first is the label
to apply, if the conditional evaluates to true.

The (optional) second one defines the name of the
label group and defaults to PRIORITY.

The resulting labels are put into the __STATIC.LABELS
section of the output, where the Python-CMake-interface
will pick them up in the future.

Check the example cond2.mini for a comprehensive test.