Make the dot dict hashable to allow discarding duplicate configs easily
The current way of constructing strings and reconstructin dotdicts from
keys was a hack to do exactly that. We do now do it via hashing. This
makes the code much easier to read and understand, it also comes at a
price though: The hash support should not be used for anything else,
especially for constructing dicts with DotDicts as keys. Mutating one
such key will make the universe collapse.