[cmake][bugfix] Trigger reconfiguration on meta ini change
This is an extremely important for the meta ini format to be
usable in practice. The technical solution is a hack, which is
very elegant nevertheless. I found it on the cmake mailing list:

http://www.cmake.org/pipermail/cmake/2013-October/055987.html

It uses the semantics of configure_file, which do automatically
trigger reconfiguration. For us reconfiguration involves reexpansion.