Integrate python docs with cmake docs in Sphinx
This turned out to be much more painfil than I thought it would.
The reason: Sphinx does not know namespace packages (quite a bummer).
As a consequence:
* We will not be able to use `sphinx-apidoc`
* We do not get rid of the errors about not finding `dune.common.x`
  (it thinks it should be in the scope of the documented module, but its not)

I have renamed the former file `dune-testtools.rst` to `python.rst`.
Use this file for the python docs and include it into the index
toctree at will.