[fuzzyvtk] Sort connectivity array cell-locally
for the comparison we don't care about the local order of the vertices
in the connectivity array. As different grid manager might produce different
local orderings, i.e. flipped or permuted versions of the same element, we
just sort them in ascending order here to make them easy to compare